feat(web): remove org inputs from reporting surfaces
