+++ apps/web/app/(dashboard)/certificates/[id]/page.tsx
@@ -1,6 +1,5 @@
 import type { Metadata } from 'next'
 import { notFound } from 'next/navigation'
-import { getRequiredSession } from '@/lib/auth/session'
 import { getCertificate } from '@/lib/actions/certificates'
 import { CertificateDetailClient } from './certificate-detail-client'
 
@@ -13,16 +12,13 @@ export default async function CertificateDetailPage({
 }: {
   params: Promise<{ id: string }>
 }) {
-  const session = await getRequiredSession()
-  const orgId = session.user.organisationId!
   const { id } = await params
 
-  const result = await getCertificate(orgId, id)
+  const result = await getCertificate(id)
   if (!result) notFound()
 
   return (
     <CertificateDetailClient
-      orgId={orgId}
       initialCertificate={result.certificate}
       initialEvents={result.events}
     />
