+++ apps/web/app/(dashboard)/directory-lookup/page.tsx
@@ -16,8 +16,7 @@ export const metadata: Metadata = {
 export default async function DirectoryLookupPage() {
   const session = await getRequiredSession()
   if (!canAccessTooling(session.user)) redirect('/dashboard')
-  const orgId = session.user.organisationId
-  const configs = orgId ? await getLookupConfigOptions(orgId) : []
+  const configs = await getLookupConfigOptions()
 
   if (configs.length === 0) {
     return (
@@ -44,5 +43,5 @@ export default async function DirectoryLookupPage() {
     )
   }
 
-  return <DirectoryLookupClient orgId={orgId ?? ''} configs={configs} />
+  return <DirectoryLookupClient configs={configs} />
 }
