+++ apps/web/app/(dashboard)/reports/patch-status/page.tsx
@@ -1,7 +1,5 @@
 import type { Metadata } from 'next'
-import { getRequiredSession } from '@/lib/auth/session'
 import { getPatchManagementReport } from '@/lib/actions/patch-status'
-import { createEmptyPatchManagementReport } from '@/lib/standalone-empty-state'
 import { PatchStatusReportClient } from './patch-status-report-client'
 
 export const metadata: Metadata = {
@@ -9,9 +7,6 @@ export const metadata: Metadata = {
 }
 
 export default async function PatchStatusReportPage() {
-  const session = await getRequiredSession()
-  const orgId = session.user.organisationId
-
-  const report = orgId ? await getPatchManagementReport(orgId) : createEmptyPatchManagementReport()
+  const report = await getPatchManagementReport()
   return <PatchStatusReportClient report={report} />
 }
