+++ apps/web/app/(dashboard)/reports/software/page.tsx
@@ -1,8 +1,6 @@
 import type { Metadata } from 'next'
 import { getRequiredSession } from '@/lib/auth/session'
-import { db } from '@/lib/db'
-import { organisations, hostGroups } from '@/lib/db/schema'
-import { eq, and, isNull } from 'drizzle-orm'
+import { listGroups } from '@/lib/actions/host-groups'
 import { SoftwareReportClient } from './software-report-client'
 import { NuqsAdapter } from 'nuqs/adapters/next/app'
 
@@ -11,34 +9,12 @@ export const metadata: Metadata = {
 }
 
 export default async function SoftwareReportPage() {
-  const session = await getRequiredSession()
-  const orgId = session.user.organisationId
-
-  if (!orgId) {
-    return (
-      <NuqsAdapter>
-        <SoftwareReportClient orgId="" orgName="CT-Ops" hostGroups={[]} />
-      </NuqsAdapter>
-    )
-  }
-
-  const [org, groups] = await Promise.all([
-    db.query.organisations.findFirst({
-      where: and(eq(organisations.id, orgId), isNull(organisations.deletedAt)),
-      columns: { id: true, name: true },
-    }),
-    db.query.hostGroups.findMany({
-      where: and(eq(hostGroups.organisationId, orgId), isNull(hostGroups.deletedAt)),
-      columns: { id: true, name: true },
-      orderBy: hostGroups.name,
-    }),
-  ])
-
-  if (!org) return null
+  await getRequiredSession()
+  const groups = await listGroups()
 
   return (
     <NuqsAdapter>
-      <SoftwareReportClient orgId={orgId} orgName={org.name} hostGroups={groups} />
+      <SoftwareReportClient orgName="CT-Ops" hostGroups={groups} />
     </NuqsAdapter>
   )
 }
