+++ apps/web/app/(dashboard)/reports/vulnerabilities/page.tsx
@@ -1,8 +1,6 @@
 import type { Metadata } from 'next'
 import { getRequiredSession } from '@/lib/auth/session'
-import { db } from '@/lib/db'
-import { hostGroups } from '@/lib/db/schema'
-import { eq, and, isNull } from 'drizzle-orm'
+import { listGroups } from '@/lib/actions/host-groups'
 import { VulnerabilityReportClient } from './vulnerability-report-client'
 
 export const metadata: Metadata = {
@@ -10,18 +8,8 @@ export const metadata: Metadata = {
 }
 
 export default async function VulnerabilityReportPage() {
-  const session = await getRequiredSession()
-  const orgId = session.user.organisationId
+  await getRequiredSession()
+  const groups = await listGroups()
 
-  if (!orgId) {
-    return <VulnerabilityReportClient orgId="" hostGroups={[]} />
-  }
-
-  const groups = await db.query.hostGroups.findMany({
-    where: and(eq(hostGroups.organisationId, orgId), isNull(hostGroups.deletedAt)),
-    columns: { id: true, name: true },
-    orderBy: hostGroups.name,
-  })
-
-  return <VulnerabilityReportClient orgId={orgId} hostGroups={groups} />
+  return <VulnerabilityReportClient hostGroups={groups} />
 }
