+++ apps/web/app/(dashboard)/reports/vulnerabilities/vulnerability-report-client.tsx
@@ -31,7 +31,6 @@ import type { VulnerabilityFindingConfidence, VulnerabilityReportFilters, Vulner
 import type { HostGroup } from '@/lib/db/schema'
 
 interface Props {
-  orgId: string
   hostGroups: Pick<HostGroup, 'id' | 'name'>[]
 }
 
@@ -47,7 +46,7 @@ function SeverityBadge({ severity }: { severity: string }) {
   return <Badge variant="outline">Unknown</Badge>
 }
 
-export function VulnerabilityReportClient({ orgId, hostGroups }: Props) {
+export function VulnerabilityReportClient({ hostGroups }: Props) {
   const [cve, setCve] = useState('')
   const [packageName, setPackageName] = useState('')
   const [severity, setSeverity] = useState<VulnerabilitySeverity | 'all'>('all')
@@ -71,8 +70,8 @@ export function VulnerabilityReportClient({ orgId, hostGroups }: Props) {
   }), [confidence, cve, distro, fixAvailable, hostGroupId, kevOnly, packageName, severity, source])
 
   const { data, isLoading, error, refetch } = useQuery({
-    queryKey: ['vulnerability-report', orgId, filters],
-    queryFn: () => getVulnerabilityReport(orgId, filters),
+    queryKey: ['vulnerability-report', filters],
+    queryFn: () => getVulnerabilityReport(filters),
     staleTime: 30_000,
   })
