+++ apps/web/app/(dashboard)/service-accounts/[id]/page.tsx
@@ -1,6 +1,5 @@
 import type { Metadata } from 'next'
 import { notFound } from 'next/navigation'
-import { getRequiredSession } from '@/lib/auth/session'
 import { getDomainAccount } from '@/lib/actions/domain-accounts'
 import { ServiceAccountDetailClient } from './service-account-detail-client'
 
@@ -13,12 +12,10 @@ export default async function ServiceAccountDetailPage({
 }: {
   params: Promise<{ id: string }>
 }) {
-  const session = await getRequiredSession()
-  const orgId = session.user.organisationId!
   const { id } = await params
 
-  const account = await getDomainAccount(orgId, id)
+  const account = await getDomainAccount(id)
   if (!account) notFound()
 
-  return <ServiceAccountDetailClient orgId={orgId} account={account} />
+  return <ServiceAccountDetailClient account={account} />
 }
