+++ apps/web/app/(dashboard)/settings/integrations/ct-cve/ct-cve-settings-client.tsx
@@ -14,7 +14,7 @@ import {
   XCircle,
 } from 'lucide-react'
 
-import { saveOrgCtCveConnectorSettings } from '@/lib/actions/ct-cve'
+import { saveCtCveConnectorSettings } from '@/lib/actions/ct-cve'
 import type { CtCveConnectorSetupOverview } from '@/lib/integrations/ct-cve/setup-status'
 import { Alert, AlertDescription, AlertTitle } from '@/components/ui/alert'
 import { Badge } from '@/components/ui/badge'
@@ -33,7 +33,6 @@ interface InitialCtCveSettings {
 }
 
 interface CtCveSettingsClientProps {
-  orgId: string
   overview: CtCveConnectorSetupOverview
   settings: InitialCtCveSettings | null
   defaults: InitialCtCveSettings
@@ -103,7 +102,6 @@ function ConnectorWarnings({ overview }: { overview: CtCveConnectorSetupOverview
 }
 
 export function CtCveSettingsClient({
-  orgId,
   overview,
   settings,
   defaults,
@@ -120,7 +118,7 @@ export function CtCveSettingsClient({
     setError(null)
     setMessage(null)
     startTransition(async () => {
-      const result = await saveOrgCtCveConnectorSettings(orgId, {
+      const result = await saveCtCveConnectorSettings({
         enabled: formData.get('enabled') === 'on',
         name: String(formData.get('name') ?? ''),
         baseUrl: String(formData.get('baseUrl') ?? ''),
@@ -149,7 +147,7 @@ export function CtCveSettingsClient({
           />
         </div>
         <p className="max-w-3xl text-sm text-muted-foreground">
-          Configure the CT-CVE connection for this organisation and review connector health.
+          Configure the CT-CVE connection for this instance and review connector health.
         </p>
       </div>
 
@@ -224,9 +222,9 @@ export function CtCveSettingsClient({
           <CardHeader className="pb-3">
             <CardTitle className="flex items-center gap-2 text-base">
               <Settings2 className="size-4 text-muted-foreground" />
-              Organisation Connection
+              Instance Connection
             </CardTitle>
-            <CardDescription>Stored in CT Ops for this organisation only.</CardDescription>
+            <CardDescription>Stored in CT Ops for this instance only.</CardDescription>
           </CardHeader>
           <CardContent>
             <form action={submitSettings} className="grid gap-4 md:grid-cols-2">
