+++ apps/web/lib/actions/ct-cve.ts
@@ -2,7 +2,9 @@
 
 import { revalidatePath } from 'next/cache'
 
+import { getRequiredSession } from '@/lib/auth/session'
 import { requireOrgAdminAccess } from '@/lib/actions/action-auth'
+import { resolveCurrentActionScope } from './action-scope'
 import {
   defaultCtCveConnectorTokenId,
   saveCtCveConnectorSettings as saveConnectorSettings,
@@ -17,27 +19,28 @@ function readString(record: Record<string, unknown>, key: string): string {
   return typeof value === 'string' ? value : ''
 }
 
-export async function saveOrgCtCveConnectorSettings(
-  orgId: string,
+export async function saveCtCveConnectorSettings(
   input: unknown,
 ): Promise<CtCveConnectorSettingsActionResult> {
+  const session = await getRequiredSession()
+  const scopeId = resolveCurrentActionScope(session)
   try {
-    await requireOrgAdminAccess(orgId)
+    await requireOrgAdminAccess(scopeId)
   } catch {
     return { error: 'You do not have permission to update CT-CVE settings' }
   }
 
   const record = input && typeof input === 'object' ? input as Record<string, unknown> : {}
   try {
-    await saveConnectorSettings(orgId, {
+    await saveConnectorSettings(scopeId, {
       enabled: record.enabled === true,
       name: readString(record, 'name'),
       baseUrl: readString(record, 'baseUrl'),
       inventoryTokenId:
-        readString(record, 'inventoryTokenId') || defaultCtCveConnectorTokenId('ctops_inventory', orgId),
+        readString(record, 'inventoryTokenId') || defaultCtCveConnectorTokenId('ctops_inventory', scopeId),
       inventoryTokenSecret: readString(record, 'inventoryTokenSecret'),
       ctCveTokenId:
-        readString(record, 'ctCveTokenId') || defaultCtCveConnectorTokenId('ctcve_findings', orgId),
+        readString(record, 'ctCveTokenId') || defaultCtCveConnectorTokenId('ctcve_findings', scopeId),
       ctCveTokenSecret: readString(record, 'ctCveTokenSecret'),
     })
     revalidatePath('/settings/integrations/ct-cve')
